Improvement in project structure, data validations, and search.

+++ PetStore/Model/CatFood.cs
@@ -8,7 +8,7 @@ namespace PetStore.Model
 {
     public class CatFood : Product
     {
-        public double WeightPounds { get; set; }
+        public decimal WeightPounds { get; set; }
         public bool KittenFood { get; set; }
     }
 }
